Rename some files
bin/CMakeLists.txt had a typo in filename, move
test_enqueue_kernel_from_binary.c from regression to runtime
