+++ tests/testsuite-regression.at
@@ -1,11 +1,6 @@
 
 AT_BANNER([Regression tests])
 
-AT_SETUP([enqueue kernel created with binary])
-AT_KEYWORDS([regression])
-AT_CHECK([POCL_WORK_GROUP_METHOD=workitemrepl $abs_top_builddir/tests/regression/test_enqueue_kernel_from_binary], 0)
-AT_CLEANUP
-
 AT_SETUP([phi nodes not replicated (repl) - lp:927573])
 AT_KEYWORDS([regression])
 AT_SKIP_IF([! grep "#define HAVE_OPENCL_HPP" $abs_top_builddir/config.h])
